+++ src/EphemeralServer.js
@@ -112,4 +112,4 @@ class EphemeralServer {
   }
 }
 
-export { EphemeralServer }
+export default EphemeralServer
