+++ test/index.spec.js
@@ -4,8 +4,7 @@
 import { expect } from 'chai'
 import sinon from 'sinon'
 import axios from 'axios'
-import EphemeralConnector from '../src/index'
-import { EphemeralServer } from '../src/server'
+import EphemeralConnector, { EphemeralServer } from '../src/index'
 import { take } from 'rxjs/operators'
 import { w3cwebsocket } from 'websocket'
